Refactor send and system events
