+++ sdk/eventgrid/azure-eventgrid/CHANGELOG.md
@@ -7,6 +7,8 @@
   - `EventGridConsumer` is now renamed to `EventGridDeserializer`.
   - `decode_cloud_event` is renamed to `deserialize_cloud_events`.
   - `decode_eventgrid_event` is renamed to `deserialize_eventgrid_events`.
+  - The system events now exist in the `azure.eventgrid.systemevents` namespace instead of `azure.eventgrid.models` namespace.
+  - The `send` method in the `EventGridPubliserClient` is now replaced by the `send_events`.
   
   **Bug Fixes**
   - `EventGridEvent` has two additional required positional parameters namely, `data` and `data_version`.
