+++ sdk/eventgrid/azure-eventgrid/samples/champion_scenarios/cs2_publish_custom_events_to_a_domain_topic.py
@@ -24,7 +24,7 @@ domain_hostname = os.environ["EG_TOPIC_HOSTNAME"]
 credential = AzureKeyCredential(domain_key)
 client = EventGridPublisherClient(domain_hostname, credential)
 
-client.send([
+client.send_events([
 	EventGridEvent(
 		topic="MyCustomDomainTopic1",
 		event_type="Contoso.Items.ItemReceived",
