+++ sdk/eventgrid/azure-eventgrid/samples/champion_scenarios/cs5_publish_events_using_cloud_events_1.0_schema.py
@@ -24,7 +24,7 @@ topic_hostname = os.environ["CLOUD_TOPIC_HOSTNAME"]
 credential = AzureKeyCredential(topic_key)
 client = EventGridPublisherClient(topic_hostname, credential)
 
-client.send([
+client.send_events([
     CloudEvent(
         type="Contoso.Items.ItemReceived",
         source="/contoso/items",
