+++ sdk/eventgrid/azure-eventgrid/samples/publish_samples/publish_with_shared_access_signature_sample.py
@@ -51,7 +51,7 @@ def publish_event():
             event_list.append(event)
 
         # publish list of events
-        client.send(event_list)
+        client.send_events(event_list)
         print("Batch of size {} published".format(len(event_list)))
         time.sleep(randint(1, 5))
