+++ sdk/eventgrid/azure-eventgrid/tests/test_eg_publisher_client.py
@@ -36,7 +36,7 @@ class EventGridPublisherClientTests(AzureMgmtTestCase):
                 event_type="Sample.EventGrid.Event",
                 data_version="2.0"
                 )
-        client.send(eg_event)
+        client.send_events(eg_event)
 
     @CachedResourceGroupPreparer(name_prefix='eventgridtest')
     @CachedEventGridTopicPreparer(name_prefix='eventgridtest')
@@ -55,7 +55,7 @@ class EventGridPublisherClientTests(AzureMgmtTestCase):
                 event_type="Sample.EventGrid.Event",
                 data_version="2.0"
                 )
-        client.send([eg_event1, eg_event2])
+        client.send_events([eg_event1, eg_event2])
 
     @CachedResourceGroupPreparer(name_prefix='eventgridtest')
     @CachedEventGridTopicPreparer(name_prefix='eventgridtest')
@@ -68,7 +68,7 @@ class EventGridPublisherClientTests(AzureMgmtTestCase):
                 event_type="Sample.EventGrid.Event",
                 data_version="2.0"
                 )
-        client.send(eg_event)
+        client.send_events(eg_event)
 
     @CachedResourceGroupPreparer(name_prefix='eventgridtest')
     @CachedEventGridTopicPreparer(name_prefix='eventgridtest')
@@ -82,7 +82,7 @@ class EventGridPublisherClientTests(AzureMgmtTestCase):
                 data_version="2.0"
                 )
         with pytest.raises(TypeError, match="Data in EventGridEvent cannot be bytes*"):
-            client.send(eg_event)
+            client.send_events(eg_event)
 
     @CachedResourceGroupPreparer(name_prefix='eventgridtest')
     @CachedEventGridTopicPreparer(name_prefix='eventgridtest')
@@ -96,7 +96,7 @@ class EventGridPublisherClientTests(AzureMgmtTestCase):
                 "data_version":"2.0"
         }
         with pytest.raises(TypeError, match="Data in EventGridEvent cannot be bytes*"):
-            client.send(eg_event)
+            client.send_events(eg_event)
 
     @CachedResourceGroupPreparer(name_prefix='eventgridtest')
     @CachedEventGridTopicPreparer(name_prefix='cloudeventgridtest')
@@ -108,7 +108,7 @@ class EventGridPublisherClientTests(AzureMgmtTestCase):
                 data = {"sample": "cloudevent"},
                 type="Sample.Cloud.Event"
                 )
-        client.send(cloud_event)
+        client.send_events(cloud_event)
 
     @CachedResourceGroupPreparer(name_prefix='eventgridtest')
     @CachedEventGridTopicPreparer(name_prefix='cloudeventgridtest')
@@ -120,7 +120,7 @@ class EventGridPublisherClientTests(AzureMgmtTestCase):
                 data = "cloudevent",
                 type="Sample.Cloud.Event"
                 )
-        client.send(cloud_event)
+        client.send_events(cloud_event)
 
     @CachedResourceGroupPreparer(name_prefix='eventgridtest')
     @CachedEventGridTopicPreparer(name_prefix='cloudeventgridtest')
@@ -132,7 +132,7 @@ class EventGridPublisherClientTests(AzureMgmtTestCase):
                 data = b"cloudevent",
                 type="Sample.Cloud.Event"
                 )
-        client.send(cloud_event)
+        client.send_events(cloud_event)
 
     @CachedResourceGroupPreparer(name_prefix='eventgridtest')
     @CachedEventGridTopicPreparer(name_prefix='cloudeventgridtest')
@@ -144,7 +144,7 @@ class EventGridPublisherClientTests(AzureMgmtTestCase):
                 data = "cloudevent",
                 type="Sample.Cloud.Event"
                 )
-        client.send([cloud_event])
+        client.send_events([cloud_event])
 
     @CachedResourceGroupPreparer(name_prefix='eventgridtest')
     @CachedEventGridTopicPreparer(name_prefix='cloudeventgridtest')
@@ -160,7 +160,7 @@ class EventGridPublisherClientTests(AzureMgmtTestCase):
                     'extension':'hello'
                     }
                 )
-        client.send([cloud_event])
+        client.send_events([cloud_event])
         internal = cloud_event._to_generated().serialize()
         assert 'reason_code' in internal
         assert 'extension' in internal
@@ -178,7 +178,7 @@ class EventGridPublisherClientTests(AzureMgmtTestCase):
                 "data": "cloudevent",
                 "type": "Sample.Cloud.Event"
         }
-        client.send(cloud_event1)
+        client.send_events(cloud_event1)
 
     @CachedResourceGroupPreparer(name_prefix='eventgridtest')
     @CachedEventGridTopicPreparer(name_prefix='eventgridtest')
@@ -193,7 +193,7 @@ class EventGridPublisherClientTests(AzureMgmtTestCase):
                 event_type="Sample.EventGrid.Event",
                 data_version="2.0"
                 )
-        client.send(eg_event)
+        client.send_events(eg_event)
 
     @CachedResourceGroupPreparer(name_prefix='eventgridtest')
     @CachedEventGridTopicPreparer(name_prefix='customeventgridtest')
@@ -210,7 +210,7 @@ class EventGridPublisherClientTests(AzureMgmtTestCase):
                     "customData": "sample data"
                     }
                 )
-        client.send(custom_event)
+        client.send_events(custom_event)
 
     @CachedResourceGroupPreparer(name_prefix='eventgridtest')
     @CachedEventGridTopicPreparer(name_prefix='customeventgridtest')
@@ -237,7 +237,7 @@ class EventGridPublisherClientTests(AzureMgmtTestCase):
                     "customData": "sample data 2"
                     }
                 )
-        client.send([custom_event1, custom_event2])
+        client.send_events([custom_event1, custom_event2])
 
     def test_send_throws_with_bad_credential(self):
         bad_credential = "I am a bad credential"
